+++ src/main/java/org/jabref/logic/importer/fileformat/PdfContentImporter.java
@@ -21,6 +21,7 @@ import org.jabref.logic.xmp.XmpUtilReader;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.LinkedFile;
 import org.jabref.model.entry.field.StandardField;
+import org.jabref.model.entry.identifier.ArXivIdentifier;
 import org.jabref.model.entry.identifier.DOI;
 import org.jabref.model.entry.types.EntryType;
 import org.jabref.model.entry.types.StandardEntryType;
@@ -244,6 +245,7 @@ public class PdfContentImporter extends PdfImporter {
         String title;
         String conference = null;
         String doi = null;
+        String arxivId = null;
         String series = null;
         String volume = null;
         String number = null;
@@ -256,6 +258,7 @@ public class PdfContentImporter extends PdfImporter {
             // special case: possibly conference as first line on the page
             extractYear();
             doi = getDoi(null);
+            arxivId = getArxivId(null);
             if (curString.contains("Conference")) {
                 fillCurStringWithNonEmptyLines();
                 conference = curString;
@@ -388,6 +391,7 @@ public class PdfContentImporter extends PdfImporter {
                 }
             } else {
                 doi = getDoi(doi);
+                arxivId = getArxivId(arxivId);
 
                 if ((publisher == null) && curString.contains("IEEE")) {
                     // IEEE has the conference things at the end
@@ -445,6 +449,9 @@ public class PdfContentImporter extends PdfImporter {
         if (doi != null) {
             entry.setField(StandardField.DOI, doi);
         }
+        if (arxivId != null) {
+            entry.setField(StandardField.ARXIVID, arxivId);
+        }
         if (series != null) {
             entry.setField(StandardField.SERIES, series);
         }
@@ -458,6 +465,9 @@ public class PdfContentImporter extends PdfImporter {
             entry.setField(StandardField.PAGES, pages);
         }
         if (year != null) {
+            if (arxivId != null) {
+                year = "20" + arxivId.substring(0, 2);
+            }
             entry.setField(StandardField.YEAR, year);
         }
         if (publisher != null) {
@@ -478,6 +488,21 @@ public class PdfContentImporter extends PdfImporter {
             }
         }
         return doi;
+    }
+
+    private String getArxivId(String arxivId) {
+        int pos;
+        if (arxivId == null) {
+            pos = curString.indexOf("arxiv");
+            if (pos < 0) {
+                pos = curString.indexOf("arXiv");
+            }
+            if (pos >= 0) {
+                String arxivText = curString.substring(pos);
+                return ArXivIdentifier.parse(arxivText).map(ArXivIdentifier::getNormalized).orElse(null);
+            }
+        }
+        return arxivId;
     }
 
     private String getFirstPageContents(PDDocument document) throws IOException {