+++ src/main/java/org/jabref/model/entry/field/StandardField.java
@@ -25,6 +25,7 @@ public enum StandardField implements Field {
     ARCHIVEPREFIX("archiveprefix"),
     ASSIGNEE("assignee", FieldProperty.PERSON_NAMES),
     AUTHOR("author", FieldProperty.PERSON_NAMES),
+    ARXIVID("arXivId", FieldProperty.VERBATIM),
     BOOKAUTHOR("bookauthor", FieldProperty.PERSON_NAMES),
     BOOKPAGINATION("bookpagination", FieldProperty.PAGINATION),
     BOOKSUBTITLE("booksubtitle", FieldProperty.BOOK_NAME),